+++ CHANGELOG.md
@@ -4,16 +4,17 @@
 
 ### Added
 
+-   [examples]: heatmap/table with custom symbol and counter
 -   [package/core]: prop `format` for `Counter` component
 -   [package/core]: prop `stretch` to the `Chart` component
 -   [examples]: examples of venn charts
 -   [package/set]: venn charts with two and three sets
--   [examples]: examples of upset charts
+-   [examples]: horizontal and vertical upset charts
 -   [package/matrix]: upset charts
--   [examples]: examples of schedule charts
+-   [examples]: horizontal and vertical schedule charts
 -   [packages/core]: tests for scales and motion
 -   [packages/band]: components `Schedule` and `Schedules` for schedule / Gantt charts
--   [examples]: examples of line charts with time scales
+-   [examples]: line charts with time scales
 -   [packages/core]: support for time scales
 
 ### Changed